feat(submissions): move ParsedResponsesObj to email-submissions.util
ParsedResponsesObject is meant to be a data class rather than a
service. Makes more sense to move it to email submissions util
together with submissionsEmailObject
