+++ src/app/modules/submission/submission.service.ts
@@ -1,4 +1,3 @@
-import _ from 'lodash'
 import mongoose from 'mongoose'
 import { err, errAsync, ok, okAsync, Result, ResultAsync } from 'neverthrow'
 
@@ -23,6 +22,7 @@ import { createQueryWithDateParam, isMalformedDate } from '../../utils/date'
 import { validateField } from '../../utils/field-validation'
 import { DatabaseError, MalformedParametersError } from '../core/core.errors'
 
+import { getFilteredResponses } from './email-submission/email-submission.util'
 import {
   ConflictError,
   ProcessingError,
@@ -30,49 +30,10 @@ import {
   ValidateFieldError,
 } from './submission.errors'
 import { ProcessedFieldResponse } from './submission.types'
-import { extractEmailConfirmationData, getModeFilter } from './submission.utils'
+import { extractEmailConfirmationData } from './submission.utils'
 
 const logger = createLoggerWithLabel(module)
 const SubmissionModel = getSubmissionModel(mongoose)
-
-/**
- * Filter allowed form field responses from given responses and return the
- * array of responses with duplicates removed.
- *
- * @param form The form document
- * @param responses the responses that corresponds to the given form
- * @returns neverthrow ok() filtered list of allowed responses with duplicates (if any) removed
- * @returns neverthrow err(ConflictError) if the given form's form field ids count do not match given responses'
- */
-const getFilteredResponses = (
-  form: IFormDocument,
-  responses: FieldResponse[],
-): Result<FieldResponse[], ConflictError> => {
-  const modeFilter = getModeFilter(form.responseMode)
-
-  if (!form.form_fields) {
-    return err(new ConflictError('Form fields are missing'))
-  }
-  // _id must be transformed to string as form response is jsonified.
-  const fieldIds = modeFilter(form.form_fields).map((field) => ({
-    _id: String(field._id),
-  }))
-  const uniqueResponses = _.uniqBy(modeFilter(responses), '_id')
-  const results = _.intersectionBy(uniqueResponses, fieldIds, '_id')
-
-  if (results.length < fieldIds.length) {
-    const onlyInForm = _.differenceBy(fieldIds, results, '_id').map(
-      ({ _id }) => _id,
-    )
-    return err(
-      new ConflictError('Some form fields are missing', {
-        formId: form._id,
-        onlyInForm,
-      }),
-    )
-  }
-  return ok(results)
-}
 
 /**
  * Injects response metadata such as the question, visibility state. In
@@ -164,112 +125,6 @@ export const getProcessedResponses = (
   }
 
   return ok(processedResponses)
-}
-
-export class ParsedResponsesObject {
-  public ndiResponses: ProcessedFieldResponse[] = []
-  constructor(public responses: ProcessedFieldResponse[]) {}
-
-  addNdiResponses(
-    ndiResponses: ProcessedFieldResponse[],
-  ): ParsedResponsesObject {
-    this.ndiResponses = ndiResponses
-    return this
-  }
-
-  /**
-   * Injects response metadata such as the question, visibility state. In
-   * addition, validation such as input validation or signature validation on
-   * verified fields are also performed on the response.
-   * @param form The form document corresponding to the responses
-   * @param responses The responses to process and validate
-   * @returns neverthrow ok() with field responses with additional metadata injected.
-   * @returns neverthrow err() if response validation fails
-   */
-  static parseResponses(
-    form: IFormDocument,
-    responses: FieldResponse[],
-  ): Result<
-    ParsedResponsesObject,
-    ProcessingError | ConflictError | ValidateFieldError
-  > {
-    const filteredResponsesResult = getFilteredResponses(form, responses)
-    if (filteredResponsesResult.isErr()) {
-      return err(filteredResponsesResult.error)
-    }
-
-    const filteredResponses = filteredResponsesResult.value
-
-    // Set of all visible fields
-    const visibleFieldIds = getVisibleFieldIds(filteredResponses, form)
-
-    // Guard against invalid form submissions that should have been prevented by
-    // logic.
-    if (
-      getLogicUnitPreventingSubmit(filteredResponses, form, visibleFieldIds)
-    ) {
-      return err(new ProcessingError('Submission prevented by form logic'))
-    }
-
-    // Create a map keyed by field._id for easier access
-
-    if (!form.form_fields) {
-      return err(new ProcessingError('Form fields are undefined'))
-    }
-
-    const fieldMap = form.form_fields.reduce<{
-      [fieldId: string]: IFieldSchema
-    }>((acc, field) => {
-      acc[field._id] = field
-      return acc
-    }, {})
-
-    // Validate each field in the form and inject metadata into the responses.
-    const processedResponses = []
-    for (const response of filteredResponses) {
-      const responseId = response._id
-      const formField = fieldMap[responseId]
-      if (!formField) {
-        return err(
-          new ProcessingError('Response ID does not match form field IDs'),
-        )
-      }
-
-      const processingResponse: ProcessedFieldResponse = {
-        ...response,
-        isVisible:
-          // Set isVisible as true for Encrypt mode if there is a response for mobile and email field
-          // Because we cannot tell if the field is unhidden by logic
-          // This prevents downstream validateField from incorrectly preventing
-          // encrypt mode submissions with responses on unhidden fields
-          // TODO(#780): Remove this once submission service is separated into
-          // Email and Encrypted services
-          form.responseMode === ResponseMode.Encrypt
-            ? 'answer' in response &&
-              typeof response.answer === 'string' &&
-              response.answer.trim() !== ''
-            : visibleFieldIds.has(responseId),
-        question: formField.getQuestion(),
-      }
-
-      if (formField.isVerifiable) {
-        processingResponse.isUserVerified = formField.isVerifiable
-      }
-
-      // Error will be returned if the processed response is not valid.
-      const validateFieldResult = validateField(
-        form._id,
-        formField,
-        processingResponse,
-      )
-      if (validateFieldResult.isErr()) {
-        return err(validateFieldResult.error)
-      }
-      processedResponses.push(processingResponse)
-    }
-
-    return ok(new ParsedResponsesObject(processedResponses))
-  }
 }
 
 /**
